+++ otree.browser.js
@@ -135,11 +135,19 @@ class ImmutableNode extends AbstractNode {
     return this._parent
   }
 
-  getPoint(relativeTo) {
+  getPoint() {
+    return this._getPoint()
+  }
+
+  _getPoint(relativeTo) {
     return {
       x: this._getXCoordinate(relativeTo),
       y: this._getYCoordinate(relativeTo)
     }
+  }
+
+  getPointRelativeTo(relativeTo) {
+    return this._getPoint(relativeTo)
   }
 
   getIndentation(relativeTo) {
@@ -149,7 +157,7 @@ class ImmutableNode extends AbstractNode {
   _getYCoordinate(relativeTo) {
     return this.isRoot(relativeTo)
       ? 0
-      : this.getRootNode(relativeTo)
+      : this._getRootNode(relativeTo)
           .getTopDownArray()
           .indexOf(this) + 1 // todo: may be slow for big trees.
   }
@@ -158,9 +166,13 @@ class ImmutableNode extends AbstractNode {
     return relativeTo === this || !this.getParent()
   }
 
-  getRootNode(relativeTo) {
+  getRootNode() {
+    return this._getRootNode()
+  }
+
+  _getRootNode(relativeTo) {
     if (this.isRoot(relativeTo)) return this
-    return this.getParent().getRootNode(relativeTo)
+    return this.getParent()._getRootNode(relativeTo)
   }
 
   toString(indentCount = 0, language = this) {
@@ -200,9 +212,17 @@ class ImmutableNode extends AbstractNode {
     return `${edgeHtml}${lineHtml}${childrenHtml}</span>`
   }
 
-  getWords(startFrom = 0) {
+  _getWords(startFrom) {
     if (!this._words) this._words = this._getLine().split(this.getZI())
     return startFrom ? this._words.slice(startFrom) : this._words
+  }
+
+  getWords() {
+    return this._getWords(0)
+  }
+
+  getWordsFrom(startFrom) {
+    return this._getWords(startFrom)
   }
 
   getKeyword() {
@@ -210,7 +230,7 @@ class ImmutableNode extends AbstractNode {
   }
 
   getBeam() {
-    const words = this.getWords(1)
+    const words = this.getWordsFrom(1)
     return words.length ? words.join(this.getZI()) : undefined
   }
 
@@ -220,15 +240,19 @@ class ImmutableNode extends AbstractNode {
     return (beam ? beam : "") + (this.length ? this.getYI() + this._childrenToString() : "")
   }
 
-  getStack(relativeTo) {
+  getStack() {
+    return this._getStack()
+  }
+
+  _getStack(relativeTo) {
     if (this.isRoot(relativeTo)) return []
     const parent = this.getParent()
     if (parent.isRoot(relativeTo)) return [this]
-    else return parent.getStack(relativeTo).concat([this])
+    else return parent._getStack(relativeTo).concat([this])
   }
 
-  getStackString(relativeTo) {
-    return this.getStack(relativeTo)
+  getStackString() {
+    return this._getStack()
       .map((node, index) => this.getXI().repeat(index) + node.getLine())
       .join(this.getYI())
   }
@@ -238,14 +262,30 @@ class ImmutableNode extends AbstractNode {
   }
 
   // todo: return array? getPathArray?
-  getKeywordPath(relativeTo) {
+  _getKeywordPath(relativeTo) {
     if (this.isRoot(relativeTo)) return ""
     else if (this.getParent().isRoot(relativeTo)) return this.getKeyword()
 
-    return this.getParent().getKeywordPath(relativeTo) + this.getXI() + this.getKeyword()
+    return this.getParent()._getKeywordPath(relativeTo) + this.getXI() + this.getKeyword()
   }
 
-  getPathVector(relativeTo) {
+  getKeywordPathRelativeTo(relativeTo) {
+    return this._getKeywordPath(relativeTo)
+  }
+
+  getKeywordPath() {
+    return this._getKeywordPath()
+  }
+
+  getPathVector() {
+    return this._getPathVector()
+  }
+
+  getPathVectorRelativeTo(relativeTo) {
+    return this._getPathVector(relativeTo)
+  }
+
+  _getPathVector(relativeTo) {
     if (this.isRoot(relativeTo)) return []
     const path = this.getParent().getPathVector(relativeTo)
     path.push(this.getIndex())
@@ -339,7 +379,7 @@ class ImmutableNode extends AbstractNode {
   }
 
   _getXCoordinate(relativeTo) {
-    return this.getStack(relativeTo).length
+    return this._getStack(relativeTo).length
   }
 
   getParentFirstArray() {
@@ -518,8 +558,14 @@ class ImmutableNode extends AbstractNode {
     return Object.keys(obj)
   }
 
-  getGraph(key) {
+  getGraphByKey(key) {
     const graph = this._getGraph(key)
+    graph.push(this)
+    return graph
+  }
+
+  getGraph() {
+    const graph = this._getGraph()
     graph.push(this)
     return graph
   }
@@ -571,7 +617,11 @@ class ImmutableNode extends AbstractNode {
     return headerRow + "\n" + rows.join("\n")
   }
 
-  toTable(maxWidth = 100, alignRight = false) {
+  toTable() {
+    return this._toTable(100, false)
+  }
+
+  toFormattedTable(maxWidth, alignRight) {
     return this._toTable(maxWidth, alignRight)
   }
 
@@ -607,12 +657,16 @@ class ImmutableNode extends AbstractNode {
     return this.toDelimited(" ")
   }
 
-  toOutline(nodeFn) {
+  toOutline() {
+    return this._toOutline(node => node.getLine())
+  }
+
+  toMappedOutline(nodeFn) {
     return this._toOutline(nodeFn)
   }
 
   // Adapted from: https://github.com/notatestuser/treeify.js
-  _toOutline(nodeFn = node => node.getLine()) {
+  _toOutline(nodeFn) {
     const growBranch = (outlineTreeNode, last, lastStates, nodeFn, callback) => {
       let lastStatesCopy = lastStates.slice(0)
       const node = outlineTreeNode.node
@@ -649,11 +703,7 @@ class ImmutableNode extends AbstractNode {
   }
 
   copyTo(node, index) {
-    const newNode = node._setLineAndChildren(
-      this.getLine(),
-      this.childrenToString(),
-      index === undefined ? node.length : index
-    )
+    const newNode = node._setLineAndChildren(this.getLine(), this.childrenToString(), index)
     return newNode
   }
 
@@ -963,7 +1013,7 @@ class TreeNode extends ImmutableNode {
     const graph = this.getGraph()
     const result = new TreeNode()
     graph.forEach(node => result.extend(node))
-    return new TreeNode().append(this.getLine(), result)
+    return new TreeNode().appendLineAndChildren(this.getLine(), result)
   }
 
   getExpanded() {
@@ -1014,7 +1064,7 @@ class TreeNode extends ImmutableNode {
     // tood: cleanup.
     const remainingString = lines.join(this.getYI())
     const children = new TreeNode(remainingString)
-    if (!remainingString) children.append("")
+    if (!remainingString) children.appendLine("")
     this.setChildren(children)
     return this
   }
@@ -1046,7 +1096,11 @@ class TreeNode extends ImmutableNode {
     return this._setChildren(tuple[1])
   }
 
-  append(line, children) {
+  appendLine(line) {
+    return this._setLineAndChildren(line)
+  }
+
+  appendLineAndChildren(line, children) {
     return this._setLineAndChildren(line, children)
   }
 
@@ -1077,7 +1131,7 @@ class TreeNode extends ImmutableNode {
   shift() {
     if (!this.length) return null
     const node = this.getChildren().shift()
-    return node.copyTo(new this.constructor())
+    return node.copyTo(new this.constructor(), 0)
   }
 
   sort(fn) {
@@ -1156,17 +1210,21 @@ class TreeNode extends ImmutableNode {
     const parent = this.getParent()
     const index = this.getIndex()
     const newNodeText = new TreeNode(str).nodeAt(0)
-    const newNode = parent.insert(newNodeText.getLine(), newNodeText.childrenToString(), index)
+    const newNode = parent.insertLineAndChildren(newNodeText.getLine(), newNodeText.childrenToString(), index)
     this.destroy()
     return newNode
   }
 
-  insert(line, children, index) {
+  insertLineAndChildren(line, children, index) {
     return this._setLineAndChildren(line, children, index)
   }
 
-  prepend(line, children) {
-    return this.insert(line, children, 0)
+  insertLine(line, index) {
+    return this._setLineAndChildren(line, undefined, index)
+  }
+
+  prependLine(line) {
+    return this.insertLine(line, 0)
   }
 
   pushBeamAndChildren(beam, children) {
@@ -1176,13 +1234,13 @@ class TreeNode extends ImmutableNode {
       index++
     }
     const line = index.toString() + (beam === undefined ? "" : this.getZI() + beam)
-    return this.append(line, children)
+    return this.appendLineAndChildren(line, children)
   }
 
   _touchNode(keywordPathArray) {
     let contextNode = this
     keywordPathArray.forEach(keyword => {
-      contextNode = contextNode.getNode(keyword) || contextNode.append(keyword)
+      contextNode = contextNode.getNode(keyword) || contextNode.appendLine(keyword)
     })
     return contextNode
   }
@@ -1218,27 +1276,36 @@ class TreeNode extends ImmutableNode {
     return this
   }
 
-  static fromCsv(str, hasHeaders) {
-    return this.fromDelimited(str, ",", hasHeaders)
+  static fromCsv(str) {
+    return this.fromDelimited(str, ",", '"')
   }
 
   static fromJson(str) {
     return new TreeNode(JSON.parse(str))
   }
 
-  static fromSsv(str, hasHeaders) {
-    return this.fromDelimited(str, " ", hasHeaders)
+  static fromSsv(str) {
+    return this.fromDelimited(str, " ", '"')
   }
 
-  static fromTsv(str, hasHeaders) {
-    return this.fromDelimited(str, "\t", hasHeaders)
+  static fromTsv(str) {
+    return this.fromDelimited(str, "\t", '"')
   }
 
-  static fromDelimited(str, delimiter, hasHeaders, quoteChar = '"') {
-    const rows = str.includes(quoteChar)
+  static fromDelimited(str, delimiter, quoteChar) {
+    const rows = this._getEscapedRows(str, delimiter, quoteChar)
+    return this._rowsToTreeNode(rows, delimiter, true)
+  }
+
+  static _getEscapedRows(str, delimiter, quoteChar) {
+    return str.includes(quoteChar)
       ? this._strToRows(str, delimiter, quoteChar)
       : str.split("\n").map(line => line.split(delimiter))
-    return this._rowsToTreeNode(rows, delimiter, hasHeaders === false ? false : true)
+  }
+
+  static fromDelimitedNoHeaders(str, delimiter, quoteChar) {
+    const rows = this._getEscapedRows(str, delimiter, quoteChar)
+    return this._rowsToTreeNode(rows, delimiter, false)
   }
 
   static _strToRows(str, delimiter, quoteChar, newLineChar = "\n") {
@@ -1390,8 +1457,9 @@ class TreeNode extends ImmutableNode {
 
         if (child.tagName && child.tagName.match(/parsererror/i)) throw new Error("Parse Error")
 
-        if (child.childNodes.length > 0 && child.tagName) children.append(child.tagName, this._treeNodeFromXml(child))
-        else if (child.tagName) children.append(child.tagName)
+        if (child.childNodes.length > 0 && child.tagName)
+          children.appendLineAndChildren(child.tagName, this._treeNodeFromXml(child))
+        else if (child.tagName) children.appendLine(child.tagName)
         else if (child.data) {
           const data = child.data.trim()
           if (data) children.pushBeamAndChildren(data)
@@ -1466,14 +1534,14 @@ class AbstractGrammarBackedProgram extends TreeNode {
     const grammarProgram = this.getGrammarProgram()
     const keywordDefinitions = grammarProgram.getKeywordDefinitions()
     keywordDefinitions.forEach(child => {
-      usage.append([child.getWord(0), "line-id", "keyword", child.getBeamParameters().join(" ")].join(" "))
+      usage.appendLine([child.getWord(0), "line-id", "keyword", child.getBeamParameters().join(" ")].join(" "))
     })
     const programNodes = this.getTopDownArray()
     programNodes.forEach((programNode, lineNumber) => {
       const def = programNode.getDefinition()
       const keyword = def.getKeyword()
       const stats = usage.getNode(keyword)
-      stats.append([filepath + "-" + lineNumber, programNode.getWords().join(" ")].join(" "))
+      stats.appendLine([filepath + "-" + lineNumber, programNode.getWords().join(" ")].join(" "))
     })
     return usage
   }
@@ -1613,7 +1681,7 @@ class GrammarConstNode extends TreeNode {
   getValue() {
     // todo: parse type
     if (this.length) return this.childrenToString()
-    return this.getWords(2).join(" ")
+    return this.getWordsFrom(2).join(" ")
   }
   getName() {
     return this.getKeyword()
@@ -1731,7 +1799,7 @@ class GrammarBackedNode extends TreeNode {
     const parameterLength = parameters.length
     const lastParameterType = parameters[parameterLength - 1]
     const lastParameterListType = lastParameterType && lastParameterType.endsWith("*") ? lastParameterType : undefined
-    const words = this.getWords(1)
+    const words = this.getWordsFrom(1)
     const length = Math.max(words.length, parameterLength)
     const checks = []
     for (let index = 0; index < length; index++) {
@@ -1883,7 +1951,7 @@ class GrammarDefinitionErrorNode extends TreeNode {
   }
 
   getWordTypeLine() {
-    return ["keyword"].concat(this.getWords(1).map(word => "any")).join(" ")
+    return ["keyword"].concat(this.getWordsFrom(1).map(word => "any")).join(" ")
   }
 }
 
@@ -2361,6 +2429,6 @@ otree.TreeNode = TreeNode
 otree.NonTerminalNode = GrammarBackedNonTerminalNode
 otree.TerminalNode = GrammarBackedTerminalNode
 
-otree.getVersion = () => "11.4.1"
+otree.getVersion = () => "11.5.0"
 
 window.otree = otree