+++ otree.d.ts
@@ -25,15 +25,19 @@ interface TreeNode {
 
   compile: (targetExtension: string) => string
   getIndex: () => int
-  getPoint: (relativeTo?: TreeNode) => point
-  getPathVector: (relativeTo?: TreeNode) => pathVector
   getLine: () => nodeString
   getChildrenByNodeType: (type: TreeNodeClass) => TreeNode[]
   getNodeByType: (type: TreeNodeClass) => TreeNode | Undefined
-  getStack: (relativeTo?: TreeNode) => TreeNode[]
-  getStackString: (relativeTo?: TreeNode) => string
+  getPoint: () => point
+  getPointRelativeTo: (relativeTo: TreeNode) => point
+  getPathVector: () => pathVector
+  getPathVectorRelativeTo: (relativeTo: TreeNode) => pathVector
+  getStack: () => TreeNode[]
+  getStackString: () => string
+  getRootNode: () => This | TreeNode
+  getKeywordPath: () => keywordPath
+  getKeywordPathRelativeTo: (relativeTo: TreeNode) => keywordPath
   getParent: () => TreeNode | undefined
-  getRootNode: (relativeTo?: TreeNode) => This | TreeNode
   getKeyword: () => word
   getExpanded: () => string
   getErrors: () => string[] // parse errors. base class is permissive and will always have 0 errors.
@@ -42,11 +46,12 @@ interface TreeNode {
   getYoungerSiblings: () => TreeNode[] // where younger sibling is a node with a higher index
   getWordTypeLine: () => string // something like "any int int". base class words are always any type.
   getWord: (index: int) => word
-  getWords: (startingFrom?: int) => word[]
+  getWords: () => word[]
+  getWordsFrom: (startingFrom: int) => word[]
   getBeam: () => string | Undefined // Always refers to part of the line after the keyword, given that ZI is space.
-  getKeywordPath: (relativeTo?: TreeNode) => keywordPath
   getTopDownArray: () => TreeNode[] // returns all nodes as array in preorder order
-  getGraph: (headKey?: word) => TreeNode[] // if no param, uses getWord(1)
+  getGraphByKey: (headKey: word) => TreeNode[]
+  getGraph: () => TreeNode[] // uses getWord(1) as the key
   getNext: () => TreeNode // wrapsaround
   getPrevious: () => TreeNode // wrapsaround
   getInheritanceTree: () => TreeNode // useful when your trees follow the convention "className parentClassName" line structure
@@ -55,7 +60,7 @@ interface TreeNode {
   isTerminal: () => Boolean
   isRoot: () => Boolean
   clone: () => TreeNode
-  copyTo: (tree: TreeNode, index?: int) => TreeNode
+  copyTo: (tree: TreeNode, index: int) => TreeNode
   getLines: () => string[]
   getNodeByColumns: (...columns: string[]) => TreeNode | Undefined
   getNode: (path: keywordPath) => TreeNode
@@ -77,10 +82,12 @@ interface TreeNode {
   toObject: () => Object
   toCsv: () => string
   toDelimited: (delimiter: string, header: word[]) => string
-  toTable: (maxWidth?: int, alignRight?: boolean) => string // Output a table with padding up to maxWidth in each cell
+  toTable: () => string // Output a table for printing
+  toFormattedTable: (maxWidth: int, alignRight: boolean) => string // Output a table with padding up to maxWidth in each cell
   toSsv: () => string
   toTsv: () => string
-  toOutline: (mapFn?: nodeMapFn) => string
+  toMappedOutline: (mapFn: nodeMapFn) => string
+  toOutline: () => string
   toString: () => string
   toXml: () => string
 
@@ -90,9 +97,10 @@ interface TreeNode {
   parseNodeType: (line: string) => TreeNode
 
   // Mutable Methods
-  append: (line: string, tree?: TreeNode) => TreeNode
+  appendLine: (line: string) => TreeNode
+  appendLineAndChildren: (line: string, tree: content) => TreeNode
   concat: (b: TreeNode | string) => This
-  delete: (path: keywordPath) => This // todo: rename delete child?
+  delete: (path: keywordPath) => This // todo: perhaps rename to delete child
   extend: (tree: TreeNode | string) => This // recursively extend the object
   destroy: () => undefined
   duplicate: () => TreeNode
@@ -100,10 +108,10 @@ interface TreeNode {
   getTreeMTime: () => number // get time tree was last modified. Initializes lazily on first call.
   setLine: (line: string) => This
   setFromText: (text: string) => This
-  insert: (line: string, tree?: TreeNode, index?: int) => TreeNode
+  insertLine: (line: string, index: int) => TreeNode
+  insertLineAndChildren: (line: string, tree: content, index: int) => TreeNode
   invert: () => This // Flips keywords and beams on all top level nodes. Does not recurse.
-  prepend: (line: string, tree?: TreeNode) => TreeNode
-  pushBeamAndTree: (beam?: string, tree?: TreeNode) => TreeNode // Keyword will be set to this.length + 1. todo: remove?
+  prependLine: (line: string) => TreeNode
   replaceNode: (fn: replaceNodeFn) => TreeNode
   remap: (key: Object) => This // Does not recurse.
   rename: (oldKeyword: word, newKeyword: word) => This
@@ -111,7 +119,7 @@ interface TreeNode {
   sortBy: (keywordOrKeywords: word | word[]) => This
   setKeyword: (keyword: word) => This
   setWord: (index: int, value: string) => This
-  setBeam: (value?: content) => This
+  setBeam: (value: content) => This
   reverse: () => This
   shift: () => TreeNode
   sort: (sortFn: sortFn) => This
@@ -126,11 +134,12 @@ interface GrammarProgram {}
 
 interface StaticTreeNode {
   nest: (lines: string, xi: int) => string // Insert lines, if any, as child nodes prefixed with the given number of XI characters
-  fromDelimited: (str: string, delimiter: string, hasHeaders?: boolean, quoteChar?: string) => TreeNode
+  fromDelimited: (str: string, delimiter: string, quoteChar: string) => TreeNode
+  fromDelimitedNoHeaders: (str: string, delimiter: string, quoteChar: string) => TreeNode
   fromJson: (str: Json) => TreeNode
-  fromCsv: (str: string, hasHeaders?: boolean) => TreeNode
-  fromSsv: (str: string, hasHeaders?: boolean) => TreeNode
-  fromTsv: (str: string, hasHeaders?: boolean) => TreeNode
+  fromCsv: (str: string) => TreeNode
+  fromSsv: (str: string) => TreeNode
+  fromTsv: (str: string) => TreeNode
   fromXml: (str: string) => TreeNode
 }