+++ src/grammar/GrammarConstNode.js
@@ -4,7 +4,7 @@ class GrammarConstNode extends TreeNode {
   getValue() {
     // todo: parse type
     if (this.length) return this.childrenToString()
-    return this.getWords(2).join(" ")
+    return this.getWordsFrom(2).join(" ")
   }
   getName() {
     return this.getKeyword()